+++ extension.js
@@ -6,7 +6,6 @@ let config
  * @param {vscode.ExtensionContext} context
  */
 function activate(context) {
-
     readConfig()
 
     vscode.workspace.onDidChangeConfiguration((e) => {
@@ -15,68 +14,30 @@ function activate(context) {
         }
     })
 
-    // delete
-    // |>............word
-    context.subscriptions.push(
-        vscode.commands.registerCommand('smart-delete.right', async () => {
-            let editor = vscode.window.activeTextEditor
-            let { document, selections } = editor
-            let maxLine = document.lineCount
-
-            if (selections.length > 1) {
-                let ranges = getRanges(sortSelections(selections).reverse(), 'right', maxLine)
-
-                for (const item of ranges) {
-                    await rightOps(editor, item)
-                }
-            } else {
-                let selection = selections[0]
-
-                await rightOps(
-                    editor,
-                    {
-                        selection: selection,
-                        range: new vscode.Range(
-                            selection.start.line,
-                            selection.start.character,
-                            maxLine,
-                            0
-                        )
-                    }
-                )
-            }
-        })
-    )
-
     // backspace
     // word............<|
     context.subscriptions.push(
         vscode.commands.registerCommand('smart-delete.left', async () => {
             let editor = vscode.window.activeTextEditor
-            let { selections } = editor
             let maxLine = 0
+            let ranges = getRanges(sortSelections(editor.selections), 'left', maxLine).reverse()
 
-            if (selections.length > 1) {
-                let ranges = getRanges(sortSelections(selections), 'left', maxLine)
+            for (const item of ranges) {
+                await leftOps(editor, item)
+            }
+        })
+    )
 
-                for (const item of ranges) {
-                    await leftOps(editor, item)
-                }
-            } else {
-                let selection = selections[0]
+    // delete
+    // |>............word
+    context.subscriptions.push(
+        vscode.commands.registerCommand('smart-delete.right', async () => {
+            let editor = vscode.window.activeTextEditor
+            let maxLine = editor.document.lineCount
+            let ranges = getRanges(sortSelections(editor.selections).reverse(), 'right', maxLine)
 
-                await leftOps(
-                    editor,
-                    {
-                        selection: selection,
-                        range: new vscode.Range(
-                            maxLine,
-                            0,
-                            selection.start.line,
-                            selection.start.character
-                        )
-                    }
-                )
+            for (const item of ranges) {
+                await rightOps(editor, item)
             }
         })
     )
@@ -86,8 +47,8 @@ async function rightOps(editor, item) {
     const { document } = editor
     const { selection, range } = item
 
-    // multi line selection
-    if (!selection.isSingleLine) {
+    // selected text
+    if (!selection.isSingleLine || !selection.active.isEqual(selection.anchor)) {
         return removeAll(editor, selection)
     }
 
@@ -98,28 +59,22 @@ async function rightOps(editor, item) {
 
     let search = await document.getText(range)
 
-    if (/^\s{2,}/.test(search)) { // multiline
-        if (!search.trim()) { // nothing but empty lines
-            await editor.edit(
-                (edit) => edit.replace(range, replace(search, /\s+/g, 'right')),
-                { undoStopBefore: false, undoStopAfter: false }
-            )
-        } else { // normal
-            await editor.edit(
-                (edit) => edit.replace(range, replace(search, /\s{2,}\S/m, 'right')),
-                { undoStopBefore: false, undoStopAfter: false }
-            )
-
-            // |>............direct next line
-            // if (!await otherLineCheck(editor, selection, 'right')) {
-            //     insertNewLine()
-            // }
+    // multiline
+    if (/^\s{2,}/.test(search)) {
+        // nothing but empty lines
+        if (!search.trim()) {
+            return changeDoc(editor, range, replaceTxt(search, /\s+/g, 'right'))
         }
-    } else if (new RegExp(`^${EOL}`).test(search)) { // end of line
-        await editor.edit(
-            (edit) => edit.replace(range, replace(search, new RegExp(EOL, 'm'), 'right')),
-            { undoStopBefore: false, undoStopAfter: false }
-        )
+        // normal
+        else {
+            await changeDoc(editor, range, replaceTxt(search, /\s{2,}\S/m, 'right'))
+
+            return formatIndent()
+        }
+    }
+    // end of line
+    else if (new RegExp(`^${EOL}`).test(search)) {
+        return changeDoc(editor, range, replaceTxt(search, new RegExp(EOL, 'm'), 'right'))
     } else {
         return removeOneChar(editor, range, 'right')
     }
@@ -129,8 +84,8 @@ async function leftOps(editor, item) {
     const { document } = editor
     const { selection, range } = item
 
-    // multi line selection
-    if (!selection.isSingleLine) {
+    // selected text
+    if (!selection.isSingleLine || !selection.active.isEqual(selection.anchor)) {
         return removeAll(editor, selection)
     }
 
@@ -142,52 +97,58 @@ async function leftOps(editor, item) {
     let search = await document.getText(range)
 
     if (/\s{2,}$/.test(search)) {
-        if (!search.trim()) { // nothing but empty lines
-            await editor.edit(
-                (edit) => edit.replace(range, replace(search, /\s+/g, 'left')),
-                { undoStopBefore: false, undoStopAfter: false }
-            )
-        } else { // normal
-            await editor.edit(
-                (edit) => edit.replace(range, replace(search, /\S\s{2,}$/g, 'left')),
-                { undoStopBefore: false, undoStopAfter: false }
-            )
+        // nothing but empty lines
+        if (!search.trim()) {
+            return changeDoc(editor, range, replaceTxt(search, /\s+/g, 'left'))
+        }
+        // normal
+        else {
+            await changeDoc(editor, range, replaceTxt(search, /\S\s{2,}$/g, 'left'))
 
-            // direct prev line............<|
-            // if (!await otherLineCheck(editor, selection, 'left')) {
-            //     insertNewLine()
-            // }
+            return formatIndent()
         }
     } else {
         return removeOneChar(editor, range, 'left')
     }
 }
 
-async function removeAll(editor, selection) {
-    await editor.edit(
-        (edit) => edit.replace(
-            new vscode.Range(selection.start, selection.end),
-            ''
-        ),
-        { undoStopBefore: false, undoStopAfter: false }
-    )
+function removeAll(editor, selection) {
+    let range = new vscode.Range(selection.start, selection.end)
+
+    return changeDoc(editor, range, '')
 }
 
-async function removeOneChar(editor, range, dir) {
+async function removeOneChar(editor, selection, dir) {
     let isLeft = dir == 'left'
+    let endChar = selection.end.character
 
-    await editor.edit(
-        (edit) => edit.replace(
-            new vscode.Range(
-                isLeft ? range.end.line : range.start.line,
-                isLeft ? range.end.character : range.start.character,
-                isLeft ? range.end.line : range.start.line,
-                isLeft ? range.end.character - 1 : range.start.character + 1
-            ),
-            ''
-        ),
-        { undoStopBefore: false, undoStopAfter: false }
+    let range = new vscode.Range(
+        isLeft ? selection.end.line : selection.start.line,
+        isLeft ? endChar : selection.start.character,
+        isLeft ? selection.end.line : selection.start.line,
+        isLeft
+            ? endChar == 0
+                ? 0
+                : endChar - 1
+            : selection.start.character + 1
     )
+
+    // direct prev line word
+    // <|word
+    if (range.isEmpty) {
+        range = new vscode.Range(
+            range.start.line - 1,
+            range.start.character,
+            range.end.line,
+            range.end.character
+        )
+        let search = await editor.document.getText(range)
+        await changeDoc(editor, range, replaceTxt(search, /\s+$/, 'left'))
+
+        return formatIndent()
+    }
+
+    return changeDoc(editor, range, '')
 }
 
 // get ranges distance for multi cursors
@@ -251,49 +212,15 @@ async function currentLineCheck(editor, cursor, dir, remove = true) {
     let check = regex.test(search)
 
     if (check && remove) {
-        await editor.edit(
-            (edit) => edit.replace(range, replace(search, regex, dir)),
-            { undoStopBefore: false, undoStopAfter: false }
-        )
+        await changeDoc(editor, range, replaceTxt(search, regex, dir))
     }
 
     return check
 }
 
-// prev/next line
-async function otherLineCheck(editor, cursor, dir) {
-    if (config.keepOneLine) {
-        const { document } = editor
-
-        // check for cursor current location
-        // ............<|word
-        // word|>............
-        if (!await currentLineCheck(editor, cursor, dir, false)) {
-            let start = cursor.start.line
-            let isLeft = dir == 'left'
-            let line = isLeft
-                ? start == 0
-                    ? 0
-                    : start - 1
-                : start + 1
-            let txt
-
-            try {
-                txt = await document.lineAt(line).text
-            } catch (err) {
-                txt = await document.lineAt(line - 1).text
-            }
-
-            return txt.trim()
-        }
-    }
-
-    return false
-}
-
 // utils
 function sortSelections(arr) {
-    return arr.sort((a, b) => { // make sure its sorted correctly
+    return arr.sort((a, b) => {
         if (a.start.line > b.start.line) return 1
         if (b.start.line > a.start.line) return -1
 
@@ -301,10 +228,23 @@ function sortSelections(arr) {
     })
 }
 
-function replace(txt, regex, dir) {
+async function formatIndent() {
+    if (config.keepOneLine && config.formatAfterNewLine) {
+        return vscode.commands.executeCommand('editor.action.formatDocument')
+    }
+}
+
+async function changeDoc(editor, range, replacement) {
+    return editor.edit(
+        (edit) => edit.replace(range, replacement),
+        { undoStopBefore: false, undoStopAfter: false }
+    )
+}
+
+function replaceTxt(txt, regex, dir) {
     let isLeft = dir == 'left'
     let space = config.keepOneLine
-        ? ''
+        ? EOL
         : config.keepOneSpace
             ? ' '
             : ''
@@ -314,12 +254,6 @@ function replace(txt, regex, dir) {
 
         return isLeft ? `${data}${space}` : `${space}${data}`
     })
-}
-
-function insertNewLine() {
-    config.keepOneLine
-        ? vscode.commands.executeCommand('type', { text: EOL })
-        : false
 }
 
 function readConfig() {